refactor(5e): Rename constants file
