+++ src/models/CharacterConstants5e.js
